put TableView Delegate and Datasource functions in MVVP+SliderView
