--- quavi/quavi/Controllers/MapViewController+UIObjects.swift
@@ -1,14 +1,0 @@
-//
-//  MapViewController+UIObjects.swift
-//  quavi
-//
-//  Created by Alex 6.1 on 2/11/20.
-//  Copyright © 2020 Sunni Tang. All rights reserved.
-//
-
-import UIKit
-
-extension MapViewController {
-    
-    
-}
